adding spring boot and maven, dir restructure
