+++ .gitignore
@@ -25,3 +25,30 @@ replay_pid*
 
 # VS code settings file
 .vscode/
+
+# Maven
+target/
+pom.xml.tag
+pom.xml.releaseBackup
+pom.xml.versionsBackup
+pom.xml.next
+release.properties
+dependency-reduced-pom.xml
+buildNumber.properties
+.mvn/timing.properties
+.mvn/wrapper/maven-wrapper.jar
+
+# IDEs
+.idea/
+*.iws
+*.iml
+*.ipr
+
+.classpath
+.project
+.settings/
+.metadata/
+
+# Local databases
+*.db
+*.sqlite
