+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.datasource.url=jdbc:sqlite:expense_tracker.db
+spring.datasource.driver-class-name=org.sqlite.JDBC
